+++ packages/ui/useScroll/useScroll.js
@@ -33,6 +33,7 @@ import css from './useScroll.module.less';
 const
 	constants = {
 		animationDuration: 1000,
+		minAnimationDuration: 100,
 		epsilon: 1,
 		flickConfig: {maxDuration: null},
 		isPageDown: is('pageDown'),
@@ -50,6 +51,7 @@ const
 	},
 	{
 		animationDuration,
+		minAnimationDuration,
 		epsilon,
 		flickConfig,
 		isPageDown,
@@ -830,6 +832,8 @@ const useScrollBase = (props) => {
 	// scrollMode 'native' ]]
 
 	function onKeyDown (ev) {
+		mutableRef.current.keyPressed = ev.repeat;
+
 		if (scrollMode === 'translate') {
 			if (props.onKeyDown) {
 				forward('onKeyDown', ev, props);
@@ -840,6 +844,11 @@ const useScrollBase = (props) => {
 			props.preventScroll?.(ev);
 			forward('onKeyDown', ev, props);
 		}
+	}
+
+	function onKeyUp (ev) {
+		mutableRef.current.keyPressed = false;
+		forward('onKeyUp', ev, props);
 	}
 
 	function scrollToAccumulatedTarget (delta, vertical, overscrollEffect) {
@@ -1190,16 +1199,24 @@ const useScrollBase = (props) => {
 		return (curTime) => {
 			const
 				{sourceX, sourceY, targetX, targetY, duration} = animationInfo,
-				bounds = getScrollBounds();
+				bounds = getScrollBounds(),
+				scrollAnimationDuration = mutableRef.current.keyPressed ? minAnimationDuration : duration;
 
-			if (curTime < duration) {
+			const dynamicScrollDuration = (target, source) => {
+				const customDuration = Math.max(minAnimationDuration, Math.abs(target - source));
+				if (mutableRef.current.keyPressed) return minAnimationDuration;
+				return animationDuration !== duration ? duration : Math.min(customDuration, duration);
+			};
+
+			if (curTime < scrollAnimationDuration) {
 				let
 					toBeContinued = false,
 					curTargetX = sourceX,
 					curTargetY = sourceY;
 
 				if (canScrollHorizontally(bounds)) {
-					curTargetX = mutableRef.current.animator.timingFunction(sourceX, targetX, duration, curTime);
+					const scrollDuration = dynamicScrollDuration(sourceX, targetX);
+					curTargetX = mutableRef.current.animator.timingFunction(sourceX, targetX, scrollDuration, curTime, mutableRef.current.keyPressed);
 
 					if (Math.abs(curTargetX - targetX) < epsilon) {
 						curTargetX = targetX;
@@ -1209,7 +1226,8 @@ const useScrollBase = (props) => {
 				}
 
 				if (canScrollVertically(bounds)) {
-					curTargetY = mutableRef.current.animator.timingFunction(sourceY, targetY, duration, curTime);
+					const scrollDuration = dynamicScrollDuration(sourceY, targetY);
+					curTargetY = mutableRef.current.animator.timingFunction(sourceY, targetY, scrollDuration, curTime, mutableRef.current.keyPressed);
 
 					if (Math.abs(curTargetY - targetY) < epsilon) {
 						curTargetY = targetY;
@@ -1483,6 +1501,7 @@ const useScrollBase = (props) => {
 	function addEventListeners () {
 		utilEvent('wheel').addEventListener(scrollContainerRef, onWheel);
 		utilEvent('keydown').addEventListener(scrollContainerRef, onKeyDown);
+		utilEvent('keyup').addEventListener(scrollContainerRef, onKeyUp);
 		utilEvent('mousedown').addEventListener(scrollContainerRef, onMouseDown);
 
 		// scrollMode 'native' [[
@@ -1504,6 +1523,7 @@ const useScrollBase = (props) => {
 	function removeEventListeners () {
 		utilEvent('wheel').removeEventListener(scrollContainerRef, onWheel);
 		utilEvent('keydown').removeEventListener(scrollContainerRef, onKeyDown);
+		utilEvent('keyup').removeEventListener(scrollContainerRef, onKeyUp);
 		utilEvent('mousedown').removeEventListener(scrollContainerRef, onMouseDown);
 
 		// scrollMode 'native' [[